refactor: move to workspace setup with subpackages for rust and python
