+++ packages/py_algos/src/py_algos/kmeans/__init__.py
@@ -1,12 +1,9 @@
 """Implementations for KMeans clustering algorithm."""
 
-from ml_practice._lowlevel import KMeansRust
-
 from .native_py import KMeans
 from .numpy_py import KMeansNumpy
 
 __all__ = [
     "KMeans",
     "KMeansNumpy",
-    "KMeansRust",
 ]
