+++ packages/py_algos/src/py_algos/linreg/__init__.py
@@ -1,12 +1,9 @@
 """Implementations for Linear Regression."""
 
-from ml_practice._lowlevel import LinRegGDRust
-
 from .native_py import LinearRegressionGD
 from .numpy_py import LinearRegressionNumpy
 
 __all__ = [
     "LinearRegressionGD",
     "LinearRegressionNumpy",
-    "LinRegGDRust",
 ]
