+++ packages/rs_algos/rs_algos.pyi
