+++ packages/rs_algos/src/linreg.rs
