+++ packages/rs_algos/src/utils.rs
