+++ pyproject.toml
@@ -5,19 +5,13 @@ description = "Add your description here"
 authors = [
     { name = "William Keating", email = "54231777+keatingw@users.noreply.github.com" },
 ]
-dependencies = [
-    "numpy>=2.0.0",
-    "scikit-learn>=1.5.1",
-    "matplotlib>=3.9.1",
-    "pandas[plot]>=2.2.2",
-    "plotly>=5.22.0",
-]
+dependencies = ["py_algos", "rs_algos"]
 readme = "README.md"
 requires-python = ">=3.11"
 
 [build-system]
-requires = ["maturin>=1.2,<2.0"]
-build-backend = "maturin"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
 [tool.uv]
 dev-dependencies = [
@@ -28,10 +22,12 @@ dev-dependencies = [
     "pytest>=8.3.2",
 ]
 
-[tool.maturin]
-python-source = "python"
-module-name = "ml_practice._lowlevel"
-features = ["pyo3/extension-module"]
+[tool.uv.sources]
+py_algos = { workspace = true }
+rs_algos = { workspace = true }
+
+[tool.uv.workspace]
+members = ["packages/py_algos", "packages/rs_algos"]
 
 [tool.ruff]
 target-version = "py312"
